threatflux_anthropic_sdk::models

Module user_profile

User Profiles API models (beta: user-profiles-2026-03-24).

Structs

EnrollmentUrl

Enrollment URL returned for a profile.

TrustGrant

UserProfile

A user profile used to attribute API requests to an end user or reseller.

UserProfileCreateRequest

Request to create a user profile.

UserProfileListParams

Parameters for listing profiles.

UserProfileUpdateRequest

Request to update a user profile. Metadata values set to an empty string are removed.

Enums

TrustGrantStatus

State of a profile trust grant.

UserProfileRelationship

Relationship between a profile and the platform owning the API key.

Type Aliases

UserProfileListResponse